% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.if.comma.or.semicolon.R
\name{detect.if.comma.or.semicolon}
\alias{detect.if.comma.or.semicolon}
\title{detect.if.comma.or.semicolon}
\usage{
detect.if.comma.or.semicolon(filepath)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
