% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.old.Data.R
\name{Display.old.Data}
\alias{Display.old.Data}
\title{Display.old.Data}
\usage{
Display.old.Data(idx)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
