% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.data.item1.R
\name{edit.data.item1}
\alias{edit.data.item1}
\title{edit.data.item1}
\usage{
\method{edit}{data.item1}(item.df, idx)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
