% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editsurvey2.R
\name{editsurvey2}
\alias{editsurvey2}
\title{checking.nominal}
\usage{
editsurvey2()
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
