% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.1.R
\name{get.1}
\alias{get.1}
\title{get.1 function}
\usage{
get.1(data, split, cols, whichcol)
}
\arguments{
\item{Click}{here and there}
}
\description{
check ordinal if they are ordinal
}
\examples{
get.1(data,split,cols,whichcol)
}
\keyword{ordinal}
