% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.1.sub.R
\name{get.1.sub}
\alias{get.1.sub}
\title{get.1.sub function}
\usage{
get.1.sub(data, split, cols, whichcol, replace)
}
\arguments{
\item{Click}{here and there}
}
\description{
check ordinal if they are ordinal
}
\examples{
get.1.sub(data,split,cols,whichcol,replace)
}
\keyword{ordinal}
