% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.element.from.list.R
\name{get.element.from.list}
\alias{get.element.from.list}
\title{get.element.from.list}
\usage{
get.element.from.list(list, el)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
