% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give.description1.R
\name{give.description1}
\alias{give.description1}
\title{give.description1}
\usage{
give.description1(id)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
