% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give.item1.R
\name{give.item1}
\alias{give.item1}
\title{give.item1}
\usage{
give.item1(idx)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
