% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.ui.R
\name{insert.ui}
\alias{insert.ui}
\title{insert.ui}
\usage{
insert.ui(item1)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
