% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.ui2.R
\name{insert.ui2}
\alias{insert.ui2}
\title{insert.ui2}
\usage{
insert.ui2()
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
