% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.csv.version.R
\name{load.csv.version}
\alias{load.csv.version}
\title{load.csv.version}
\usage{
load.csv.version()
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
