% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.var.index.R
\name{make.var.index}
\alias{make.var.index}
\title{make.var.index}
\usage{
make.var.index()
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
