% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.csvX.R
\name{read.csvX}
\alias{read.csvX}
\title{read.csvX}
\usage{
read.csvX(filepath)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
