% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.ui2.R
\name{remove.ui2}
\alias{remove.ui2}
\title{remove.ui2}
\usage{
remove.ui2()
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
