% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.code1_old.R
\name{save.code1_old}
\alias{save.code1_old}
\title{save.code1}
\usage{
save.code1_old(c1, id)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
