% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.csv.version.R
\name{save.csv.version}
\alias{save.csv.version}
\title{save.csv.version}
\usage{
save.csv.version(dat)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
