% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.entry.item1.R
\name{save.entry.item1}
\alias{save.entry.item1}
\title{save.entry.item1}
\usage{
save.entry.item1(item1, input)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
