% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.entry.vldes.R
\name{save.entry.vldes}
\alias{save.entry.vldes}
\title{save.entry.vldes}
\usage{
save.entry.vldes(id, input)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
