% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.form.R
\name{show.form}
\alias{show.form}
\title{show.form}
\usage{
show.form(onlyshow = NULL)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
