% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.fun.R
\name{test.fun}
\alias{test.fun}
\title{test.fun}
\usage{
test.fun(ip)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
