% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textInputRow.R
\name{textInputRow}
\alias{textInputRow}
\title{textInputRow}
\usage{
textInputRow(inputId, label, value = "")
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
