% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateInputs.R
\name{UpdateInputs}
\alias{UpdateInputs}
\title{UpdateInputs}
\usage{
UpdateInputs(idx, session)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
